rename lift to cli
