+++ tests/test_cli.py
@@ -3,11 +3,11 @@
 '''
 test_lift.py
 
-A module that contains tests for the lift.py module
+A module that contains tests for the cli.py module
 '''
 
 import unittest
-from forklift import lift
+from forklift import cli
 from json import loads
 from os import remove
 from os.path import abspath, dirname, join, exists
@@ -26,7 +26,7 @@ class TestLift(unittest.TestCase):
             remove('config.json')
 
     def test_init_creates_config_file(self):
-        path = lift.init()
+        path = cli.init()
 
         self.assertTrue(exists(path))
 
@@ -35,42 +35,45 @@ class TestLift(unittest.TestCase):
 
     def test_list_pallets(self):
         test_pallets_folder = join(self.test_data_folder, 'list_pallets')
-        pallets = lift.list_pallets(paths=[test_pallets_folder])
+        pallets = cli.list_pallets(folders=[test_pallets_folder])
 
         self.assertEquals(len(pallets), 3)
         self.assertEquals(pallets[0][0], join(test_pallets_folder, 'multiple_pallets.py'))
         self.assertEquals(pallets[0][1], 'PalletOne')
 
     def test_set_config_paths_requires_list(self):
-        self.assertRaises(Exception, lift._set_config_paths, 'hello')
+        self.assertRaises(Exception, cli._set_config_folders, 'hello')
 
     def test_add_pallet_folder(self):
-        path = lift.init()
+        path = cli.init()
 
-        lift.add_pallet_folder('another/folder')
+        cli.add_config_folder(abspath('tests\data'))
 
         with open(path) as config:
-            self.assertEquals(['c:\\scheduled', 'another/folder'], loads(config.read()))
+            self.assertEquals(['c:\\scheduled', abspath('tests\data')], loads(config.read()))
 
     def test_add_pallet_folder_checks_for_duplicates(self):
-        lift.init()
+        path = cli.init()
 
-        lift.add_pallet_folder('another/folder')
-        self.assertRaises(Exception, lift.add_pallet_folder, 'another/folder')
+        cli.add_config_folder(abspath('tests\data'))
+        cli.add_config_folder(abspath('tests\data'))
+
+        with open(path) as config:
+            self.assertEquals(['c:\\scheduled', abspath('tests\data')], loads(config.read()))
 
     def test_remove_pallet_folder(self):
-        path = lift.init()
+        path = cli.init()
         test_config_path = join(self.test_data_folder, 'remove_test_config.json')
 
         with open(path, 'w') as json_data_file, open(test_config_path) as test_config_file:
             json_data_file.write(test_config_file.read())
 
-        lift.remove_pallet_folder('path/one')
+        cli.remove_pallet_folder('path/one')
 
         with open(path) as test_config_file:
             self.assertEquals(['path/two'], loads(test_config_file.read()))
 
     def test_remove_pallet_folder_checks_for_existing(self):
-        lift.init()
+        cli.init()
 
-        self.assertRaises(Exception, lift.remove_pallet_folder, 'blah')
+        self.assertEquals('{} is not in the config folders list!'.format('blah'), cli.remove_pallet_folder('blah'))
